Made ScEd compile more pleasantly
git-svn-id: http://svn.wildfiregames.com/public/ps/trunk@1700 3db68df2-c116-0410-a063-a993310a9797

+++ source/tools/sced/InfoBox.cpp
@@ -95,7 +95,7 @@ void render(COverlayText* overlaytext)
 	if (!font) return;
 
 	const CStr& str=overlaytext->GetString();
-	int len=str.Length();
+	int len=(int)str.Length();
 	if (len==0) return;
 
 	// bind to font texture
